calculate.rbを削除し、計算料金はsimulator.rbに記述
