+++ challengeA/yoshiharu_chiba/lib/json_plan.json
@@ -2,23 +2,40 @@
   {
     "provider_name": "東京電力エナジーパートナー",
     "plan_name": "従量電灯Bプラン",
-    "price": ""
+    "price": "",
+    "basic_price": {
+      "10": 286,
+      "15": 429,
+      "20": 572,
+      "30": 858,
+      "40": 1144,
+      "50": 1430,
+      "60": 1716
+    }
   },
   {
     "provider_name": "Looopでんき",
     "plan_name": "おうちプラン",
-    "price": ""
+    "price": "",
+    "basic_price": {
+      "10": 0,
+      "15": 0,
+      "20": 0,
+      "30": 0,
+      "40": 0,
+      "50": 0,
+      "60": 0
+    }
   },
   {
     "provider_name": "東京ガス",
     "plan_name": "ずっとも電気１プラン",
-    "price": ""
+    "price": "",
+    "basic_price": {
+      "30": 858,
+      "40": 1144,
+      "50": 1430,
+      "60": 1716
+    }
   }
 ]
-
-// 新規電力会社は末尾に以下を追記する
-// {
-//   "provider_name": "テスト",
-//   "plan_name": "テストプラン",
-//   "price": ""
-// }
